+++ test/unittest/TArgvContext.cpp
@@ -1,7 +1,7 @@
 #include "ArgvContext.hpp"
 #include "gtest/gtest.h"
 
-using namespace Limonp;
+using namespace limonp;
 
 TEST(ArgvContextTest, Test1) {
   const char * argv[] = {"./exe1", "--hehe", "11", "key2", "-k", "val"};