+++ ChangeLog.md
@@ -1,5 +1,9 @@
 # limonp ChangeLog
 
+## next version
+
++ `namespace limonp`, not `Limonp` .
+
 ## 0.1.0.
 
 + Basic functions